backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2026.2
